Factor out county wizard ballot audit stage.
